refactor: 컴포넌트 디렉토리 기능별 분리
