+++ FE/pages/profile.js
@@ -3,8 +3,8 @@ import { useEffect } from "react";
 import Router from "next/router";
 import Layout from "../components/Layout";
 import Head from "next/head";
-import UsernameEditForm from "../components/UsernameEditForm";
-import FollowList from "../components/FollowList";
+import UsernameEditForm from "../components/my/UsernameEditForm";
+import FollowList from "../components/my/FollowList";
 
 const Profile = () => {
   const me = useSelector((state) => state.user.me);
